Begin Reorganizing How Tabs Work
The global "highlight" or "use tags" settings are a bit redundant, since
you can already disable highlighting all together by unchecking the
boxes next to those settings in the options page.  I'd like tags to work
the same way.  The setting for the "number of friends who want this" tag
is basically what I'm after.  I'd like individual tags that can be
turned off and re-colored to the user's preference.  For this particular
code, I've only implemented the "Owned" tag this way, but I'd eventually
like all tags to follow suit and remove the global "highlight with tags"
setting.