Update appid caching (#525)
This update makes it so that a user's GetOwnedGames list is only pulled
once every 7 days (rather than every hour).  It also changes the
behavior of on_app_info so that it will only pull get_app_details on
apps that aren't already marked as owned.

The end result is that Enhanced Steam will essentially be compiling a
list of owned appIDs which are only flushed when the user logs out or
manually clears their cache.  Again, this will significantly cut down on
the number of appuserdetails API calls as well as cut down on the hourly
reliance on the GetOwnedGames API.