Brought back pack highlighting
...but with a bit of a twist.

Now, all of the appids for a package will be batched together to make a
single API call to appuserdetails.  In addition to the value of the
subid itself being cached, the individual returned values for each appid
are also cached.

This allows Enhanced Steam to continue highlighting packages with a very
low impact on the number of API calls (one per sub, per user, per hour).