Remove timing restrictions on get_app_details
This function is now only used for ajax data that is loaded onto the
page dynamically, and even then it only queries appuserdetails when the
data isn't cached.

Additionally, it's only looking for the value of "is_owned", which in
testing has always returned the correct value even if the API is being
throttled.