Fix a bug with wishlist loading
If a user isn't logged in to the community site, the /my url redirects
to a login page.  This way, it will always resolve to a user's wishlist,
even if that user is not logged in.  However, if a user's profile is
private and they're not logged in, they'll still have problems.