REMOVE: local appid database (Fixes #525)
Since content on Steam is now appropriately labeled as "In library" and
"on wishlist", Enhanced Steam no longer needs to maintain a list of apps
which are in a user's library or are on the user's wishlist.  This
commit fundamentally changes how highlighting is done with Enhanced
Steam, in that now it simply looks for the .ds_owned_flag or
.ds_wishlist_flag selectors and highlights accordingly.  Other functions
which used to rely on the local database have been further simplified to
take advantage of the new store's capabilities which has eliminated the
need for the ensure_appid_deferred function.  This change should
effectively eliminate the Steam API being queried at all to determine if
a user owns a particular app or not, thus resolving Issue #525.