UPDATE: new translation scheme
Each language will now have it's own JSON file which will be loaded on
an as-needed basis.  This should increase the extension's responsiveness
and reduce the memory footprint (since only the translations that are
needed will be loaded into memory)

Currently, this breaks functionality with the language warning feature,
but I'll get this fixed before the next release.