ADD: Basic support for HTTPS
For now, this will add basic HTTPS support to Enhanced Steam.  There are
still several factors involved that could cause this to malfunction or
break SSL integrity, most of which are issues on Valve's end that
Enhanced Steam won't be able to fix.

HTTPS Everywhere (and similar extensions) should work correctly now
without loading insecure content, but some content (such as requests to
the ES API servers, for now) will fail to respond on HTTPS.