UPDATE: Total Spent feature for new page layout
This is pretty rough, as Valve didn't clearly mark or label any of these
divs like they used to, thus we have to resort to using text-based
pattern recognition.  So for the moment, this will only work for users
in English.  Additionally, this will currently only pull the most recent
2,000 consolidated transactions from the user's history so that will
also need updated going forward.

But it's a start.