UPDATE: non-async loop of account history
This should now correctly loop completely through the user's account
history, 1,000 rows at a time.  The process is called asynchronously so
that it can be evaluated and tallied before moving on to the next XHR
call.  For the vast majority, this should be an extremely low
performance impact.  Anyone else can simply disable that feature as
needed.