FIX: regression of currency detection
Sadly the appdetails API won't return the correct currency code unless
?cc= is specified.  Also sadly, Enhanced Steam pulls this code from a
cookie value that may or may not be correct for all users.  As such, we
are reverting to loading the store page for Half Life 2 and parsing the
user's currency value from the price field, which will reliably return
the user's currency value on the store.