Unified promises requesting for dynamicstore
* Instead of caching parts of data from
store.steampowered.com/dynamicstore/userdata/ we now cache it whole and
pick what we need from it later. This results in less code and less
requests when needing that data
* The owned apps ids required for "Launch Random Game" feature are also
pulled from this data now instead of calling the ES API

To ensure the cache is even fresher we could add an option to force
updates when actions are taken. Such as wishlist changes, ignoring apps,
buying apps etc.