Fix page hanging while scrolling in search
Fix a bug that would cause slow scrolling in search
While in search with endless scrolling enabled the page hangs while
scrolling. Chances are this is not so noticeable on faster machines but
is quite noticeable on a laptop with an i5 processor.

Here are two videos with the issue, before and after:
https://streamable.com/je60
https://streamable.com/4dvx