UPDATE: signed_in_promise
Apparently, users with private profiles that request /profiles/0/ will
not return g_steamID.  So unfortunately we have to revert this back to
loading the user's own profile.

I've added additional sanity checks to the code that should now prevent
it from throwing errors and halting execution if it doesn't find exactly
what it's looking for.