add_wishlist_total() optimizations
When accessing the Wishlist, having many wishlisted apps can make the
page freeze or stutter for a few moments, which sucks since it looks as
it's done loading and ready to interact with. So, if there are too many
items, we display a button which will load the wishlist total when/if
needed.