Tags improvements and optimizations
*I messed up by forgetting to switch the branch before starting these
changes so I'll have to commit them in this PR, sorry about that.*

- tags are inserted faster and in a cleaner way
- changed hightlights default colors to go better with the overall site
design
- changed tags design to go better with the overall site design
- changed the way the background colors are applied on highlighted
content to look better and be more consistend. Instead of fading from
complete black to the highlight color now it looks like it fades from a
darker nuance of the highlight color
- added an option to display short tags
- added tags to a few more blocks
- made optimizations to the Early Access banner in tabbed apps to be
consistent across all tabs
- renamed WebKitMutationObserver instances to MutationObserver

Note: I've added a comment for the selectors I could test if the tags
are added properly and look as expected. Others I couldn't test are also
marked.