Fix language reset
There seems to be some changes to the way the language is set. A simple
cookie change on steamcommunity.com no longer works and there is a
request to http://steamcommunity.com/actions/SetLanguage/ involved.
This is needed most likely because now the language selection is shared
with the Steam app.

To fix this we'll simply use Steam's method for changing the language.