Progress error log
When an error occurs a log of what request failed can be viewed by
hovering the progress bar.
By default, the url requested and the error code and message will be
shown, but custom error messages can also be set. There is commented out
code for testing purposes, please remove it when done with it.