ADD: Version update dialog
The idea here is that the value of the version variable (at the top of
enhancedsteam.js) gets pushed into the user's chrome.storage.sync and is
subsequently checked on every page load to see if the current running
version is different from the value in the sync storage.  If it is
different, this means ES has been updated.  If the sync storage value is
empty, this would imply that ES has been freshly installed, and we might
be able to expand this more in the future with a "Thank you for
installing Enhanced Steam" dialog, possibly with a feature tour.

TODO: message passing back to the extension's context when the user
clicks on "Don't show again" to actually disable these messages.  We
also need to add a configuration for this setting in the options page,
in case a user toggles it on accident or changes their mind about seeing
update notifications.

Also of note: the changelog_new.html file will need to be updated with
that version's changelog (in HTML format) before that version is built
and pushed to the users.