FIX: "not well-formed" Firefox error
Firefox was throwing a "not well-formed" error into the console when
loading strings.json, because the mimeType wasn't specified when
retrieving data from the local source.  Although the error didn't
prevent execution, this cleans up the console and makes sure it is done
"correctly".  Does not appear to have any adverse effects in Chrome, as
it's now just technically "more correct".