FIX: Multiple spaces between Quick and Instant sell buttons
A check was readded in order to fix this bug and another one where
buttons with wrong prices would be shown for a few moments.
The bug manifests when clicking multiple items in inventory before one
or more had a chance to finish loading the pricing information. And when
done loading the information is inserted without checking whether it
belongs to the item currently selected.