FIX: issue with quicksell functions
The issue was that the eventlistener was being added on each rebind,
causing multiple event listeners to fire resulting in additional ajax
sell requests.  The event listener for these functions has been combined
with the existing event listener in inventory_market_prepare() to
prevent this.  I sold about 100 items with heavy debugging in place, and
did not once encounter an issue using this new method.