ADD: steam review rating to wishlist
This data should probably be cached userside in a future update.  It
would also be good to add a sorting option for this as well (best rated
games to the top).  Additionally, we may look into distinguishing the
subtle "Overwhelmingly positive", "Mostly positive", etc although it
would need to be further localized.