FIX: customize homepage issues with sidebar
Valve used to push the content to the right to offset the page at higher
resolutions, which required the (now deleted) code to push everything
back to the center of the page when the sidebar was hidden.  Now, the
page's content is already centered and the sidebar is shown to the left,
so we can simply hide it like the other elements rather than shifting
the page around.