UPDATE: wishlist review cache on non-hit
This will cache the games that don't hit the API for 1 day also.  At
first this might be somewhat of an issue (as the review cache is being
built) but after a few days it will be a total non-issue and should lead
to significantly less empty-return API calls going forward.