FIX: Some issues with wishlist reviews
For reasons I'm not quite aware of, moving the build_reviews() section
below the runInPageContext section caused the tooltips to not be in
place (even though they were present before it's called).  This will
also fix the bug where reviews weren't being built on the page when they
were retrieved from the API (only from cache).