UPDATE: Protocol handling in ES
Previously we used the // convention which works perfectly fine in
Chrome but has a huge number of issues in Firefox and Edge.  To simplify
the codebase for all three builds, I've decided to implement a global
protocol variable that checks the window.location.protocol and prepends
that to locations that previously just had //.  This should fix a ton of
numerous Firefox and Edge bugs while maintaining compatibility with
Chrome.