ADD: custom profile links
This will allow users to configure a custom site to show on user
profiles.  This will help facilitate smaller sites being able to
leverage this functionality with their users without us having to make
any global changes.  The default custom site will create a link to
Google the user's Steam ID and is given as an example of the expected
syntax.  If no [ID] is configured in the URL, the code will
automatically append it to the end of the URL.  User inputs are
sanitized before injection.

Ideally in the future this could be expanded to allow for the
configuration of multiple custom sites.