REMOVE: relist function
Unfortunately, Valve made a change where items returned to inventory
from sale have a regenerated assetid.  This means that Enhanced Steam is
not able to capture the assetid programmatically like it did in the
past.  While it might be possible to reload the user's inventory and
pull the relevant ID, this could potentially lead to a type mismatch
that causes a different item to be sold from the user's inventory at an
incorrect price.  Because of this (for now at least) I'll simply be
removing the feature until a point in time where it can be brought back
reliably.