ADD: ability to activate multiple keys at once
This feature needs a bit more bug testing but the basics are in place.
Currenty my account is locked out from ativating keys for who knows how
long as I attempted to get this just right.

Valve's ajaxregister API endpoint returns a lot of undocumented status
codes that we'll probably want to attempt to document as much as
possible so we can inform the user exactly what happened with their key
activation.  Before being locked out though I was able to get this to
work reliably so I'll try another round of testing in a few hours and
adjsut the code to suit.

Additionally, we will probably want to validate / sanitize the user's
input to match Steam's key patterns and warn the user if we know their
request is going to fail.