Changed quicksell to Fetch API
This should remediate jshackles/Enhanced_Steam#1494 if the $sideMarketActs.append() lines that are commented out in the Firefox repo around 5260 are re-enabled. I don't know what the original bug was, testing with the current jshackles/Enhanced_Steam_Firefox repo showed $.ajax() wasn't sending Referer: or Origin: headers and consequently Steam was replying with a 400 Bad Request response (or without the necessary CORS headers). Changing to Fetch and setting the headers now receives the usual mix of 200 and 502 ("There was an error listing your item, try again later.") responses.

The Fetch API is available in Chrome 42+, Firefox 39+ and Edge 14+ (https://caniuse.com/#feat=fetch)