UPDATE: jQuery to 3.3.1
This update is being required by Mozilla, but I'm happy to oblige considering most of the project was semantically coded correctly to begin with (mostly thanks to @JoiNNN).

Additionally, Valve retired their old tooltip binding code and have replaced it with something easier and more consistant.  Unfortunately that means a few features are failing to load in 9.8, and this should correct that.