Rewrote Currency.promise() to set customCurrency & storeCurrency
Previously you had to manually get the currencies from the current page.
This is now done in the promise so that callers can easily access the customCurrency and storeCurrency.
This also resolves the confusion between the wallet/store currency and the converted/custom currency.

Updated the Price constructor and Price.parseFromString accordingly as well as occurences of Currency.userCurrency