Remove download button for images
Can't implement this function without adding the downloads permission as a required permission, which isn't worth it.

The problem is that the request is cross-origin, which results in just opening the image in a new tab due to https://developers.google.com/web/updates/2018/02/chrome-65-deprecations#block_cross-origin_a_download.