Replace Price.parseFromString with Price constructor
parseFromString assumed that the user wants to directly convert a string price value to a Price object with a (possibly) different currency.

But all use cases of this function were to simply create a Price object from a string, not having the intention to convert to any other currency.

The problem of this was that ambigous currency symbols (namely JPN and CNY) were just interpreted as JPY, resulting in wrong interpretations.

Fixes #656