Change to persistent background page
We have a lot of use cases for the webRequest API and our workarounds to using this API despite running an event page started to create some trouble.
For whatever reason requesting the webRequest permissions dynamically would (on some Edge browsers) grant access, but finally not trigger registered callbacks on the event page.
I could consistently reproduce this behavior by removing the extension, reopening the browser and installing the extension again.

Fixes #1043