Fix undefined notes on Firefox
IDB returns `undefined` if an entry doesn't exist.
This value was directly returned to the `sendMessage` listener.
Due to different implementations of the browser API though (might also be the polyfill, but I couldn't find anything that would change the value there) Chrome converts `undefined` to `null`.
The fix is for Firefox, but only because the assumptions for the return value were based on Chrome's implementation.

Fixes #1221